+++ lib/chid.rb
@@ -1,8 +1,6 @@
-require 'rake'
 require 'yaml'
 require 'tty-prompt'
 require 'http'
-require 'easy_translate'
 
 require 'chid/command'
 
@@ -11,71 +9,7 @@ dir = File.join(File.dirname(__FILE__))
 glob = Dir.glob(dir + "/chid/**/*.rb")
 glob.each { |r| require r }
 
-dir_tasks = File.expand_path('..', File.dirname(__FILE__))
-dir = File.join(dir_tasks, 'tasks')
-glob = Dir.glob(dir + "/**/*.rake")
-glob.each { |r| load r }
-
 module Chid
-
-  def self.chid_config
-    @chid_config ||= ChidConfig.new
-  end
-
-  def self.start
-    prompt = TTY::Prompt.new(help_color: :green)
-    confirm_install = -> (action, &block) {
-      matched = /^install:(.*)/.match(action)
-      return block.() unless matched
-
-      action_name = matched.captures.first
-      if  prompt.yes?("Can I install the #{action_name}?")
-        block.()
-      else
-        puts "\nNo problem. What do you need?"
-      end
-    }
-
-    Main.new(chid_config).init do |action, args|
-      rake_task = Rake::Task[action]
-      task_args = Rake::TaskArguments.new(rake_task.arg_names, args)
-
-      confirm_install.(action) do
-        rake_task.execute(task_args)
-      end
-
-      puts "\nDone! What else?"
-    end
-  end
-
-  def self.chid_config_path
-    home_path = File.expand_path("~/")
-    chid_config_path = File.join(home_path, '.chid.config')
-    chid_config_path
-  end
-
-  def self.username
-    on_linux { return  %x[echo $USER].strip }
-    on_osx   { return  %x[echo $(logname)].strip }
-  end
-
-  def self.on_linux
-    if  platform =~ /Linux/
-      yield
-    end
-  end
-
-  def self.on_osx
-    if  platform =~ /Darwin/
-      yield
-    end
-  end
-
-  def self.platform
-    %x{echo $(uname -s)}.strip
-  end
-
-
   # The Regex Actions are used to execute automatically some Rake::Task
   #
   # The Keys are the name of the task