+++ lib/chid/commands/installs/node.rb
@@ -22,11 +22,11 @@ Usage:
         def run
           puts "\nInstalling the Node..."
 
-          ::Chid::on_linux do
+          ::ChidConfig.on_linux do
             system('sudo apt-get install nodejs')
           end
 
-          ::Chid::on_osx do
+          ::ChidConfig.on_osx do
             system('brew install node')
           end