+++ lib/chid/commands/installs/rvm.rb
@@ -23,14 +23,14 @@ Usage:
         def run
           puts "\nInstalling the RVM..."
 
-          ::Chid::on_linux do
+          ::ChidConfig.on_linux do
             system('sudo apt-add-repository -y ppa:rael-gc/rvm')
             system('sudo apt-get update')
             system('sudo apt-get install curl')
             system('sudo apt-get install rvm')
           end
 
-          ::Chid::on_osx do
+          ::ChidConfig.on_osx do
             system('\curl -sSL https://get.rvm.io | bash')
           end